Only resize default values upon request in Coupleable
We identified the culprit for the large rise in memory usage post
PR #15790. It's the AD default members. For each qp, you add 51 Reals
worth of memory. So now we only resize default values when they are
actually requested by consumers. In this way customers like Griffin
which don't use AD at all will not incur that associated memory expense.