I changed the rank four tensor function calls to take a templated vector or array and to also take references closes #16858
Some of the rank four tensor functions were constructing a vector in order to call the fill method.
Daniel helped me template this to take an initializer_list as well.

I also switched some of the function call signatures to take references since AD Reals will be passed into them

I switched all of the internal fill methods to call the fillSymmetric21 method since it should be a little cheaper.

I switched computeVariableIsotropicElasticityTensor to call the symmetric rank four tensor since that it does not
need to call the antisymmetric tensor fill method.