Geochemistry module code-coverage improvements
This commit focusses on:
GeochemicalDatabaseValidator
GeochemistryIonicStrength
GeochemistrySpeciesSwapper

Note that in GeochemistrySpeciesSwapper I've just commented-out sections of the code that weren't tested.  I want to retain the commented-out bits because i'm not 100% sure of the formulation.  Viz, i don't see why the free electron couldn't be swapped into the basis, but i haven't found another geochem code that allows this, and conceptually perhaps it's a little strange because it suggests a high concentration of electrons floating around in solution.  So, at present, the commented-out parts are impossible to reach (so needn't be tested) but in the future someone might want that code, and it's a bit mind boggling, so i've decided to keep it.

Refs #17134