rdg: fix rdgP0 test
Previously we were evaluating the discontinuous "tracer" field at the
nodes. This does not make sense since it is CONSTANT, MONOMIAL field
and therefore multiply-defined at the nodes. The issue was only
uncovered by using a different PointLocator implementation, that
happened to give the value from the other neighboring element.  I
"fixed" the issue by having the evaluation points be slightly to the
right of each node point (except for the left and right endpoints)
instead.

Refs libMesh/libmesh#2865