Regold failing contact tests after changing primary node projection
We changed the projection method for primary nodes in a previous commit.
More specifically we now use the same normal field for projecting
primary nodes as we do for projecting secondary nodes, with the normal
field corresponding to interpolation of secondary nodal normals. This
change was made with the target of creating a consistent mortar mesh,
e.g as an example we should never run into the case where a primary node
projects onto a secondary node, but the secondary node doesn't project
onto the primary node (or visa versa). We ran into this issue quite a
bit in a fairly simple input that Antonio sent me, but the change to a
single normal projection field fixed the issue (or made it a lot better
at a minimum).

With the projection method changed, we have to regold some contact tests