Prevent strong chance of failed physical point inversion
Right now when someone calls `FEProblemBase::reinitElemPhys` with a
vector of physical points, we will attempt to invert those physical
points into the reference element space for both undisplaced and
displaced meshes if the latter exists. It is not difficult to imagine
that for a given set of physical points you may easily fall off the
displaced mesh for large enough displacements/deformations. The same
issues exists for `FEProblemBase::reinitNeighborPhys`. Callers of these
phys methods should call through the appropriate `SubProblem` that is
relevant to their context. Moreover, we should always make sure that the
caller is calling with an element corresponding to the `SubProblem`
they've selected.

Closes #17355