Add Parameter to Enable Final Jacobian Testing
The new parameter in `PetscJacobianTester.py` allows for testing of only
the final Jacobian comparison. The default behavior is only changed when
the new parameter is used.

Added a tensor mechanics test using the new parameter. If the new
parameter is removed, this new test would fail due to other Jacobian
comparisons being above the tolerance ratio.

Closes #17456