Implement node loop in ComputeWeightedGapLMMechanicalContact
This node loop occurs after we have performed the mortar segment element
loop. This implementation keeps the added LOC at a minimum and prevents
duplicate work compared to an implementation in which we loop over nodes
and then loop over elements to compute the weighted gap. I tried to
create a number of methods in `ComputeWeightedGapLMMechanicalContact`
that can hopefully make follow-on frictional work quite easy. For
instance I envision that Antonio would make `onNode` virtual as well as
potentially `computeQpProperties, computeQpIProperties` and then
override in a frictional class (potentially calling down to the base
class methods in his override).

Closes #17481